#### 1.2.1:  Updated browser-tests
Updated the browser-tests in
test_browser.html to work with
this latest version of CISF.

CISF requires an ES6-compliant
JavaScript engine. Latest versions
of Chrome, FireFox and Edge
were tested to work.  IE-11 was
tested not to work.

What you CAN do for older browsers
is to replace the CISF with a
trivial version  whose API functions
are named like the real  ones but
which do nothing, check nothing.

CISF is mostly a development-time
mechanism anyway