2.1: path(), fs()
Two new api-functions 'path' and
'fs' now come along with require("cisf")
if you are running on Node.js.

Their values are the normal 'path'
and 'fs' a normal require() gives
you on Node.js.

The reason they are now part of
Cisf is that this spares you from
having to require them yourself,
they "hitch a ride" with Cisf: