v. 4.0.2: Refactored Type-Tests
Refactored the test-function 'Type()'
into multiple inner functions one per
the type of Type being tested.

If you view the code in an IDE that
allows you to see a listing of
functions in the file this helps
when you want to look up the
behavior and thus the de facto
definition  of a given type of type.

KNOWN ISSUE:
Do not use  w(RegExp) on Edge-browser.
The tests for w(RegExpo are disabled

when ruiinning on a browser. See the
code in test.js.